split the enum file out into seperate enums
