+++ Banking/src/banking_dev/AccountType.java
@@ -1,0 +1,3 @@
+package banking_dev;
+
+enum AccountType { SAVINGS, CHECKING }
