+++ Banking/src/banking_dev/EmployeeType.java
@@ -1,0 +1,3 @@
+package banking_dev;
+
+enum EmployeeType { EMPLOYEE, SUPERVISOR }
